merch request was fixed
