+++ frontend/src/components/AddToCartButton/AddToCartButton.scss
@@ -35,14 +35,4 @@
     background-color: $bordo;
     pointer-events: none;
   }
-
-  &.added-to-cart {
-    
-
-    &:hover {
-      // color: $bordo;
-      // background-color: $main-background;
-      // border-color: $bordo;
-    }
-  }
 }
