+++ frontend/src/components/AddToFavButton/AddToFavButton.scss
@@ -12,7 +12,4 @@
     width: 46px;
     height: 46px;
   }
-
-  &:hover {
-  }
 }
