+++ frontend/src/components/AddressModal/AddressModal.scss
@@ -64,7 +64,6 @@
     }
   }
 
-
   &__input {
     width: 544px;
     height: 56px;
