+++ frontend/src/components/Cart/Cart.scss
@@ -1,6 +1,5 @@
 @import "../../styles/main.scss";
 
-
 .Cart {
   width: 512px;
   position: absolute;
