--- frontend/src/components/Dropdown/Dropdown.scss
@@ -1,85 +1,0 @@
-@import "../../styles/main.scss";
-
-.Dropdown {
-  position: relative;
-
-  &__triger {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 8px;
-    padding: 12px;
-    @extend %body;
-    color: $black;
-    width: 544px;
-    height: 56px;
-    outline: none;
-    border: 1px solid $grey;
-
-    &:focus {
-      border-color: $black;
-    }
-
-    @include onBigDesktop {
-      width: 424px;
-      height: 48px;
-    }
-
-  }
-
-  &__select {
-    position: absolute;
-    top: 99%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    background-color: $main-background;
-    border: 1px solid $black;
-    max-height: 336px;
-    overflow-y: auto;
-    z-index: 1;
-
-    &::-webkit-scrollbar {
-      width: 8px;
-    }
-    
-    &::-webkit-scrollbar-track {
-      box-shadow: inset 0 0 5px $beige; 
-      border-radius: 8px;
-    }
-     
-    &::-webkit-scrollbar-thumb {
-      background: rgba(128, 128, 128, 0.1); 
-      border-radius: 10px;
-    }
-    
-    &::-webkit-scrollbar-thumb:hover {
-      cursor: pointer;
-      background: rgba(128, 128, 128, 0.3);        
-    }
-
-
-    @include onBigDesktop {
-      max-height: 144px;
-    }
-
-    &-option {
-      height: 56px;
-      padding-left: 12px;
-      line-height: 56px;
-      @extend %body;
-      color: $grey;
-      text-align: center;
-
-      @include onBigDesktop {
-        height: 48px;
-        line-height: 48px;
-      }
-
-      &:hover {
-        background-color: $beige;
-        color: $black;
-      }
-    }
-  }
-}
