--- frontend/src/components/Dropdown/index.tsx
@@ -1,1 +1,0 @@
-export * from './Dropdown';
