+++ frontend/src/components/Footer/Footer.tsx
@@ -1,17 +1,17 @@
 import { NavLink } from 'react-router-dom';
 import './Footer.scss';
 import { getNavClassName } from '../../helpers/getNavClassName';
+import { INSTA_LINK } from '../../contants/others';
+import React from 'react';
 
-export const Footer = () => {
-
-
+export const Footer = React.memo(() => {
   return (
     <footer className="Footer">
       <div className="Footer__container container">
         <div className="Footer__nav">
           <div className="Footer__nav-item">
             <a 
-              href="https://www.instagram.com/dryg.brand/?utm_source=ig_web_button_share_sheet&igshid=OGQ5ZDc2ODk2ZA==" 
+              href={INSTA_LINK} 
               target="_blank"
               rel="noreferrer" 
             >
@@ -35,9 +35,7 @@ export const Footer = () => {
             >
               exchange and returns
             </NavLink>            
-          </div>
-
-    
+          </div>    
         </div>
 
         <div className="Footer__rights">
@@ -48,4 +46,4 @@ export const Footer = () => {
       </div>
     </footer>
   );
-};
+});
