+++ frontend/src/components/Loader/Loader.scss
@@ -14,7 +14,7 @@
     height: 2em;
     margin: 1em auto;
     border: 0.2em solid #dddddd8b;
-    border-left-color: #1f1f1f;
+    border-left-color: $black;
     animation: load8 1.2s infinite linear;
   }
 }
