+++ frontend/src/components/LoginForm/LoginForm.tsx
@@ -11,14 +11,19 @@ interface FormValues {
   password: string;
 }
 
-
 export const LoginForm = () => {
   const { userLogin, setIsLoginModalOpen } = useContext(AuthContext);
   const {pathname} = useLocation();
   const navigate = useNavigate();
-  const initialValues: FormValues = { email: '', password: '' };
+  const initialValues: FormValues = { 
+    email: '', 
+    password: '', 
+  };
 
-  const handleLoginClick = (values: FormValues, action: FormikHelpers<FormValues>) => {
+  const handleLoginClick = (
+    values: FormValues, 
+    action: FormikHelpers<FormValues>
+  ) => {
     userLogin(values)
       .then(() => {
         if (pathname === '/account/login')  {
@@ -61,7 +66,7 @@ export const LoginForm = () => {
               placeholder="Email"
               autoComplete="username"
               className={classNames('Form__field', {
-                'is-error': errors.email && touched.email
+                'is-error': errors.email && touched.email,
               })}
               onChange={handleChange}
               onBlur={handleBlur}
@@ -69,19 +74,20 @@ export const LoginForm = () => {
               validate={validateEmail}
             />
             {errors.email && touched.email && (
-              <div className="Form__error-message">{errors.email}</div>
+              <div className="Form__error-message">
+                {errors.email}
+              </div>
             )}
           </div>
 
           <div className="Form__container">
-
             <Field
               type="password"
               name="password"
               placeholder="Password"
               autoComplete="current-password"
               className={classNames('Form__field', {
-                'is-error': errors.password && touched.password
+                'is-error': errors.password && touched.password,
               })}
               onChange={handleChange}
               onBlur={handleBlur}
@@ -90,7 +96,9 @@ export const LoginForm = () => {
               
             />
             {errors.password && touched.password && (
-              <div className="Form__error-message">{errors.password}</div>
+              <div className="Form__error-message">
+                {errors.password}
+              </div>
             )}
           </div>
