+++ frontend/src/components/Logo/Logo.tsx
@@ -1,7 +1,8 @@
+import React from 'react';
 import './Logo.scss';
 
-export const Logo = () => (
+export const Logo = React.memo(() => (
   <a href="/" className="Logo">
     <div className="Logo__img" />
   </a>
-)
+));
