+++ frontend/src/styles/animations.scss
@@ -20,9 +20,11 @@
 @keyframes fade-in-left {
   0% {
     transform: translateX(120px);
+    opacity: 0;
   }
   100% {
     transform: translateX(0);
+    opacity: 1;
   }
 }
 
@@ -47,4 +49,15 @@
 
 @keyframes slide {
   100% { left: 0; opacity: 1; }
+}
+
+@keyframes fade-in-up-s {
+  0% {
+    opacity: 0;
+    transform: translateY(-10px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
 }
