+++ frontend/src/styles/blocks/Form.scss
@@ -3,6 +3,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  position: relative;
 
   &__container {
     position: relative;
@@ -58,6 +59,23 @@
       align-self: flex-start;
       position: absolute;
       bottom: -14px;
+    }
+
+    &-limit {
+      background-color: $error-background;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      border: 1px solid $bordo;
+      height: 32px;
+      width: 100%;
+      font-size: 12px;
+      color: $bordo;
+      align-self: flex-start;
+      position: absolute;
+      bottom: -48px;
+      text-align: center;
+      animation: fade-in-up-s 0.3s ease-in-out;
     }
   }
